+++ config/routes.rb
@@ -78,15 +78,18 @@ Rails.application.routes.draw do
       resources :email_read_status, only: %i[update], param: :email_id
     end
     resources :organisations, only: %i[index]
+    resources :establishments, only: %i[index]
+    resources :establishment_groups, only: %i[index]
     resources :cases, only: %i[index show edit update new create] do
       resources :interactions, only: %i[new create show]
       scope module: :cases do
         collection do
           resource :preview, only: %i[new create], as: :create_case_preview
         end
+        resource :merge_emails, only: %i[new create show], path: "merge-emails"
         resource :organisation, only: %i[edit update]
         resource :contact_details, only: %i[edit update]
-        resource :categorisation, only: %i[edit update]
+        resource :categorisation, only: %i[edit update destroy]
         resource :savings, only: %i[edit update]
         resource :procurement_details, only: %i[edit update]
         resources :documents, only: %i[show]