+++ db/migrate/20220207141256_create_support_establishment_searches.rb
@@ -1,0 +1,5 @@
+class CreateSupportEstablishmentSearches < ActiveRecord::Migration[6.1]
+  def change
+    create_view :support_establishment_searches
+  end
+end