+++ spec/factories/user.rb
@@ -54,7 +54,7 @@ FactoryBot.define do
     trait :one_supported_school do
       orgs do
         [{
-          "urn": "urn-type-1",
+          "urn": "100253",
           "name": "Specialist School for Testing",
           "type": { "id": ORG_TYPE_IDS.first, "name": "Community School" },
         }]
@@ -65,14 +65,51 @@ FactoryBot.define do
       orgs do
         [
           {
-            "urn": "urn-type-1",
+            "urn": "100253",
             "name": "Specialist School for Testing",
             "type": { "id": ORG_TYPE_IDS.first, "name": "Community School" },
           },
           {
-            "urn": "greendale-urn",
+            "urn": "100254",
             "name": "Greendale Academy for Bright Sparks",
             "type": { "id": ORG_TYPE_IDS.last, "name": "Academy Special Converter" },
+          },
+        ]
+      end
+    end
+
+    trait :one_supported_group do
+      orgs do
+        [{
+          "uid": "2314",
+          "name": "Testing Multi Academy Trust",
+          "category": { "id": GROUP_CATEGORY_IDS.first, "name": "Multi-academy Trust" },
+        }]
+      end
+    end
+
+    trait :many_supported_schools_and_groups do
+      orgs do
+        [
+          {
+            "urn": "100253",
+            "name": "Specialist School for Testing",
+            "type": { "id": ORG_TYPE_IDS.first, "name": "Community School" },
+          },
+          {
+            "urn": "100254",
+            "name": "Greendale Academy for Bright Sparks",
+            "type": { "id": ORG_TYPE_IDS.last, "name": "Academy Special Converter" },
+          },
+          {
+            "uid": "2314",
+            "name": "Testing Multi Academy Trust",
+            "category": { "id": GROUP_CATEGORY_IDS.first, "name": "Multi-academy Trust" },
+          },
+          {
+            "uid": "2315",
+            "name": "New Academy Trust",
+            "category": { "id": GROUP_CATEGORY_IDS.last, "name": "Single-academy Trust" },
           },
         ]
       end