+++ spec/features/self-serve/faf/create_framework_request_spec.rb
@@ -1,8 +1,16 @@
 RSpec.feature "Create a new framework request" do
-  xcontext "when user with one supported school" do
+  context "when user with one supported school", js: true do
     let(:user) { create(:user, :one_supported_school, first_name: "Generic", last_name: "User", full_name: "Generic User") }
 
-    xcontext "when user is already signed in" do
+    before do
+      create(:support_organisation, urn: "100253", name: "School #1")
+      stub_request(:post, "https://api.notifications.service.gov.uk/v2/notifications/email")
+      .to_return(body: {}.to_json, status: 200, headers: {})
+    end
+
+    context "when user is already signed in" do
+      let(:answers) { find_all("dd.govuk-summary-list__value") }
+
       before do
         user_is_signed_in(user: user)
         visit "/procurement-support"
@@ -13,21 +21,23 @@ RSpec.feature "Create a new framework request" do
       end
 
       it "skips user straight to how can we help page" do
-        click_start
+        find("a", text: "Start now").click
         expect(find("label.govuk-label--l")).to have_text "How can we help?"
         expect(page).to have_current_path "/procurement-support/new"
       end
 
       it "navigates to check answers page from how we can help" do
-        click_start
+        find("a", text: "Start now").click
         fill_in "framework_support_form[message_body]", with: "I have a problem"
         click_continue
-        expect(all("dd.govuk-summary-list__value")[3]).to have_text "I have a problem"
-        expect(FrameworkRequest.count).to eq(1)
-        expect(find("h1.govuk-heading-l")).to have_text "Send your request"
+        expect(answers[4]).to have_text "I have a problem"
       end
 
       it "navigates to confirmation page upon sending request" do
+        find("a", text: "Start now").click
+        fill_in "framework_support_form[message_body]", with: "I have a problem"
+        click_continue
+        click_button "Send request"
         expect(find("h1.govuk-panel__title")).to have_text "Your request for support has been sent"
       end
     end
@@ -73,9 +83,8 @@ RSpec.feature "Create a new framework request" do
           expect(page).to have_current_path "/procurement-support"
         end
 
-        # not h1
         it "loads the page" do
-          expect(find("h1")).to have_text "Do you have a DfE Sign-in account linked to the school that your request is about?"
+          expect(page).to have_text "Do you have a DfE Sign-in account linked to the school that your request is about?"
         end
 
         it "errors if no selection is given" do
@@ -94,44 +103,42 @@ RSpec.feature "Create a new framework request" do
         end
       end
 
-      describe "contact info page" do
+      describe "contact info page", js: true do
         before do
+          user_exists_in_dfe_sign_in(user: user)
           visit "/procurement-support/new"
           choose "Yes, use my DfE Sign-in"
           click_continue
-          user_exists_in_dfe_sign_in(user: user)
-          click_on "Sign in"
         end
 
-        it "loads the page" do
-          pp page.source
+        it "displays the correct attributes on the page" do
           expect(find("span.govuk-caption-l")).to have_text "About you"
           expect(find("h1.govuk-heading-l")).to have_text "Is this your contact information?"
-          expect(all("p")[1]).to have_text "If these details are not correct, you can either:"
+          expect(all("p")[1]).to have_text "These are the details we have for you, if they are not correct or up to date you will need to either:"
 
           within("ul.govuk-list") do
-            expect(all("li")[0]).to have_text "log in with the correct account, or"
-            expect(all("li")[1]).to have_link "amend your DfE Sign-in account details", href: "https://test-profile.signin.education.gov.uk/edit-details", class: "govuk-link"
+            expect(all("li")[0]).to have_text "log back in with the correct account"
+            expect(all("li")[1]).to have_link "update your DfE Sign In account details", href: "https://test-profile.signin.education.gov.uk/edit-details", class: "govuk-link"
           end
 
           within("dl.govuk-summary-list") do
-            expect(all("dt")[0]).to have_text "Name"
-            expect(all("dd")[0]).to have_text "first_name last_name"
-            expect(all("dt")[1]).to have_text "Email address"
+            expect(all("dt")[0]).to have_text "Your name"
+            expect(all("dd")[0]).to have_text "Generic User"
+            expect(all("dt")[1]).to have_text "Your email address"
             expect(all("dd")[1]).to have_text "test@test"
           end
 
-          expect(page).to have_button "Yes, continue"
+          expect(find("a.govuk-button")).to have_text "Yes, continue"
         end
       end
 
       describe "user organisation page" do
         before do
+          user_exists_in_dfe_sign_in(user: user)
           visit "/procurement-support/new"
           choose "Yes, use my DfE Sign-in"
           click_continue
-          user_exists_in_dfe_sign_in(user: user)
-          click_on "Sign in"
+          click_continue
         end
 
         it "skips step 3 because the school is implicit" do
@@ -140,26 +147,25 @@ RSpec.feature "Create a new framework request" do
         end
       end
 
-      xdescribe "user query page" do
+      describe "user query page" do
         before do
-          user_is_signed_in(user: user)
+          user_exists_in_dfe_sign_in(user: user)
           visit "/procurement-support/new"
           find("label", text: "Yes, use my DfE Sign-in").click
-          click_continue
           click_continue
           click_continue
         end
 
         it "has a back link to step 3" do
           click_on "Back"
-          expect(page).to have_current_path "/procurement-support?faf_form%5Bback%5D=true&faf_form%5Bdsi%5D=true&faf_form%5Bmessage_body%5D=&faf_form%5Bschool_urn%5D=urn-type-1&faf_form%5Bstep%5D=4"
+          expect(page).to have_current_path "/procurement-support"
         end
 
         it "raises a validation error if no message entered" do
           click_continue
 
           expect(find("h2.govuk-error-summary__title")).to have_text "There is a problem"
-          expect(page).to have_link "You must tell us how we can help", href: "#faf-form-message-body-field-error"
+          expect(page).to have_link "You must tell us how we can help", href: "#framework-support-form-message-body-field-error"
           expect(find(".govuk-error-message")).to have_text "You must tell us how we can help"
         end
 
@@ -171,20 +177,14 @@ RSpec.feature "Create a new framework request" do
         end
       end
 
-      xdescribe "check answers page" do
+      describe "check answers page" do
         before do
-          create(:support_organisation, urn: "urn-type-1", name: "School #1")
-
-          user_is_signed_in(user: user)
-          # start DSI journey
+          user_exists_in_dfe_sign_in(user: user)
           visit "/procurement-support/new"
-          # step 1
-          choose "Yes, use my DfE Sign-in"
+          find("label", text: "Yes, use my DfE Sign-in").click
           click_continue
-          # step 2
-          click_on "Yes, continue"
-          # step 4
-          fill_in "faf_form[message_body]", with: "I have a problem"
+          click_continue
+          fill_in "framework_support_form[message_body]", with: "I have a problem"
           click_continue
         end
 
@@ -192,7 +192,7 @@ RSpec.feature "Create a new framework request" do
           expect(find("h1.govuk-heading-l")).to have_text "Send your request"
           within("dl.govuk-summary-list") do
             expect(all("dt.govuk-summary-list__key")[0]).to have_text "Your name"
-            expect(all("dd.govuk-summary-list__value")[0]).to have_text "first_name last_name"
+            expect(all("dd.govuk-summary-list__value")[0]).to have_text "Generic User"
             expect(all("dd.govuk-summary-list__actions")[0]).not_to have_link "Change"
 
             expect(all("dt.govuk-summary-list__key")[1]).to have_text "Your email address"
@@ -203,9 +203,13 @@ RSpec.feature "Create a new framework request" do
             expect(all("dd.govuk-summary-list__value")[2]).to have_text "School #1"
             expect(all("dd.govuk-summary-list__actions")[2]).not_to have_link "Change"
 
-            expect(all("dt.govuk-summary-list__key")[3]).to have_text "Description of request"
-            expect(all("dd.govuk-summary-list__value")[3]).to have_text "I have a problem"
-            expect(all("dd.govuk-summary-list__actions")[3]).to have_link "Change"
+            expect(all("dt.govuk-summary-list__key")[3]).to have_text "School type"
+            expect(all("dd.govuk-summary-list__value")[3]).to have_text "Single"
+            expect(all("dd.govuk-summary-list__actions")[3]).not_to have_link "Change"
+
+            expect(all("dt.govuk-summary-list__key")[4]).to have_text "Description of request"
+            expect(all("dd.govuk-summary-list__value")[4]).to have_text "I have a problem"
+            expect(all("dd.govuk-summary-list__actions")[4]).to have_link "Change"
           end
           expect(find("p.govuk-body")).to have_text "Once you send this request, we will review it and get in touch within 2 working days."
           expect(page).to have_button "Send request"
@@ -214,11 +218,69 @@ RSpec.feature "Create a new framework request" do
     end
   end
 
-  # xcontext "when user with many supported schools" do
-  #   xcontext "and user is already signed in" do
-  #   end
+  context "when user with many supported schools", js: true do
+    let(:user) { create(:user, :many_supported_schools_and_groups, first_name: "Generic", last_name: "User", full_name: "Generic User") }
 
-  #   xcontext "and is not already signed in" do
-  #   end
-  # end
+    before do
+      create(:support_organisation, urn: "100253", name: "Specialist School for Testing")
+      create(:support_organisation, :with_address, urn: "100254", name: "Greendale Academy for Bright Sparks", phase: 7, number: "334", ukprn: "4346", establishment_type: create(:support_establishment_type, name: "Community school"))
+      create(:support_establishment_group, :with_address, uid: "2314", name: "Testing Multi Academy Trust", establishment_group_type: create(:support_establishment_group_type, name: "Multi-academy Trust"))
+      create(:support_establishment_group, :with_address, uid: "2315", name: "New Academy Trust", establishment_group_type: create(:support_establishment_group_type, name: "Single-academy Trust"))
+    end
+
+    context "and user is already signed in" do
+      before do
+        user_is_signed_in(user: user)
+        visit "/procurement-support"
+        find("a", text: "Start now").click
+      end
+
+      describe "user organisation page" do
+        it "lists all supported schools" do
+          expect(find("span.govuk-caption-l")).to have_text "About your school"
+          expect(find("legend.govuk-fieldset__legend")).to have_text "Which school are you buying for?"
+          expect(page).to have_unchecked_field "Specialist School for Testing"
+          expect(page).to have_unchecked_field "Greendale Academy for Bright Sparks"
+          expect(page).to have_unchecked_field "Testing Multi Academy Trust (MAT)"
+          expect(page).to have_unchecked_field "New Academy Trust (MAT)"
+        end
+      end
+
+      describe "check answers page" do
+        before do
+          choose "Greendale Academy for Bright Sparks"
+          click_continue
+          fill_in "framework_support_form[message_body]", with: "I have a problem"
+          click_continue
+        end
+
+        it "shows the CYA page" do
+          expect(find("h1.govuk-heading-l")).to have_text "Send your request"
+          within("dl.govuk-summary-list") do
+            expect(all("dt.govuk-summary-list__key")[0]).to have_text "Your name"
+            expect(all("dd.govuk-summary-list__value")[0]).to have_text "Generic User"
+            expect(all("dd.govuk-summary-list__actions")[0]).not_to have_link "Change"
+
+            expect(all("dt.govuk-summary-list__key")[1]).to have_text "Your email address"
+            expect(all("dd.govuk-summary-list__value")[1]).to have_text "test@test"
+            expect(all("dd.govuk-summary-list__actions")[1]).not_to have_link "Change"
+
+            expect(all("dt.govuk-summary-list__key")[2]).to have_text "Your school"
+            expect(all("dd.govuk-summary-list__value")[2]).to have_text "Greendale Academy for Bright Sparks"
+            expect(all("dd.govuk-summary-list__actions")[2]).to have_link "Change"
+
+            expect(all("dt.govuk-summary-list__key")[3]).to have_text "School type"
+            expect(all("dd.govuk-summary-list__value")[3]).to have_text "Single"
+            expect(all("dd.govuk-summary-list__actions")[3]).not_to have_link "Change"
+
+            expect(all("dt.govuk-summary-list__key")[4]).to have_text "Description of request"
+            expect(all("dd.govuk-summary-list__value")[4]).to have_text "I have a problem"
+            expect(all("dd.govuk-summary-list__actions")[4]).to have_link "Change"
+          end
+          expect(find("p.govuk-body")).to have_text "Once you send this request, we will review it and get in touch within 2 working days."
+          expect(page).to have_button "Send request"
+        end
+      end
+    end
+  end
 end