+++ spec/models/support/interaction_spec.rb
@@ -11,9 +11,9 @@ RSpec.describe Support::Interaction, type: :model do
     expect(interaction.case.request_text).to eq "This is an example request for support - please help!"
   end
 
-  it "can be a note, email (inbound/outbound) or phone call" do
+  it "can be a note, email (inbound/outbound), phone call, framework request, procurement details update, existing contract update, new contract update or savings update" do
     expect(interaction).to define_enum_for(:event_type).with_values(%i[
-      note phone_call email_from_school email_to_school support_request hub_notes hub_progress_notes hub_migration faf_support_request
+      note phone_call email_from_school email_to_school support_request hub_notes hub_progress_notes hub_migration faf_support_request procurement_updated existing_contract_updated new_contract_updated savings_updated
     ])
   end