+++ CHANGELOG.md
@@ -16,8 +16,14 @@ The format is based on [Keep a Changelog 1.0.0].
 
 ## [unreleased]
 
--
-
+### "Find a Framework" support for trusts
+- form steps reordered [#ref](https://github.com/DFE-Digital/buy-for-your-school/pull/866)
+- add 'Is this the group or trust?' page [#ref](https://github.com/DFE-Digital/buy-for-your-school/pull/847)
+- add trusts to DSI journey and update kick-out page [#ref](https://github.com/DFE-Digital/buy-for-your-school/pull/855)
+- add school type to the CYA page [#ref](https://github.com/DFE-Digital/buy-for-your-school/pull/864)
+- add group search page [#ref](https://github.com/DFE-Digital/buy-for-your-school/pull/850)
+- add 'Is it a school or group?' page [#ref](https://github.com/DFE-Digital/buy-for-your-school/pull/849)
+- make trust members eligible to access the service [#ref](https://github.com/DFE-Digital/buy-for-your-school/pull/863)
 ## [release-019] - 2022-01-31
 
 - add redirect to case management at login for agents [#ref](https://github.com/DFE-Digital/buy-for-your-school/pull/803)